requirements.txt
attrs==17.4.0
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
enum34==1.1.6
funcsigs==1.0.2
idna==2.6
lxml==4.1.1
namedlist==1.7
pluggy==0.6.0
py==1.5.2
pytest==3.4.1
pytest-allure-adaptor==1.7.10
requests==2.18.4
six==1.11.0
urllib3==1.22
